--- routes/todosRoute.js
@@ -1,8 +1,0 @@
-const express = require('express');
-
-router = express.Router();
-todosRoute = require('../controllers/todos');
-
-router.get('/', todosRoute.todos);
-
-module.exports = router;
